requirements.txt
blinker==1.9.0
click==8.2.0
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
pytz==2025.2
Werkzeug==3.1.3
